댓글 작성, 삭제 시 권한 체크 (테스트 못해봄)

+++ src/main/java/com/skrookies/dahaezlge/controller/qna/QnaController.java
@@ -523,6 +523,19 @@ public class QnaController {
     @RequestMapping("/qnaReplyDelete")
     public String qnaReplyDelete(@RequestParam("qna_re_id") Long qna_re_id, @RequestParam("qna_id") Long qna_id, HttpSession session) {
         try {
+            String userId = (String) session.getAttribute("user_id");
+            Integer userLevel = (Integer) session.getAttribute("user_level");
+            QnaDto qnaDetail = QnaService.getQnaById(Math.toIntExact(qna_id));
+
+            if (userId == null || userId.isEmpty()) {
+                // 사용자 인증 실패 시 로그인 페이지로 리다이렉트
+                return "redirect:/loginForm";
+            }
+            if((userLevel != 123 && !Objects.equals(qnaDetail.getQna_user_id(), userId))) {
+                session.setAttribute("errorMessage", "권한이 없는 사용자입니다.");
+                return "redirect:/qnaList";
+            }
+
             // 해당 답글을 삭제
             QnaService.deleteById(qna_re_id);
             session.setAttribute("messagedeleteReply", "댓글이 삭제되었습니다.");
@@ -540,10 +553,17 @@ public class QnaController {
     public String qnaReplyProcess(Model model,HttpSession session, @RequestParam("qna_id") int qna_id, @RequestParam("qna_re_body") String qna_re_body) {
         // 세션에서 user_id 확인
         String userId = (String) session.getAttribute("user_id");
+        Integer userLevel = (Integer) session.getAttribute("user_level");
+        QnaDto qnaDetail = QnaService.getQnaById(qna_id);
 
         if (userId == null || userId.isEmpty()) {
             // 세션에 user_id가 없으면 로그인 페이지로 리다이렉트
             return "redirect:/loginForm";
+        }
+
+        if((userLevel != 123 && !Objects.equals(qnaDetail.getQna_user_id(), userId))) {
+            session.setAttribute("errorMessage", "권한이 없는 사용자입니다.");
+            return "redirect:/qnaList";
         }
 
         // 댓글 내용이 비어 있는지 먼저 확인
